Added sessid requirement to update and deleteAll

+++ BackEnd/app/src/main/java/org/campusmarket/app/controllers/UserController.java
@@ -1,10 +1,13 @@
 package org.campusmarket.app.controllers;
 
+import java.util.ArrayList;
 import java.util.List;
 
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
+import org.campusmarket.app.models.Session;
 import org.campusmarket.app.models.User;
+import org.campusmarket.db.repositories.SessionsRepository;
 import org.campusmarket.db.repositories.UsersRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -14,7 +17,7 @@ import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
-import org.springframework.web.bind.annotation.ResponseStatus;
+import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.server.ResponseStatusException;
 
@@ -27,6 +30,9 @@ public class UserController
 {
     @Autowired
     private UsersRepository users;
+
+    @Autowired
+    private SessionsRepository sessions;
 
 
     Log log = LogFactory.getLog(UserController.class);
@@ -119,38 +125,70 @@ public class UserController
     }
     
     @RequestMapping(value = "/delete/all", method = RequestMethod.DELETE)
-    public void deleteAll()
+    public void deleteAll(@RequestParam(name = "sessid", required = true) String sessid)
     {
-        try
+        if (sessid.isEmpty())
         {
-            users.deleteAll();
-            log.info("User Table Cleared: all users removed.");
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Request Invalid: Empty value for required parameter 'sessid'.");
         }
-        catch(Exception e)
+
+        Session active = sessions.findBySessId(sessid);
+        
+        if (active == null) throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Could not find an active session with id: " + sessid);
+
+        if (active.getUser().getAdmin())
         {
-            log.error(e.getMessage());
-            throw new ResponseStatusException(HttpStatus.NO_CONTENT, "No users in database to remove.");
+            try
+            {
+                users.deleteAll();
+                log.info("User Table Cleared: all users removed.");
+            }
+            catch(Exception e)
+            {
+                log.error(e.getMessage());
+                throw new ResponseStatusException(HttpStatus.NO_CONTENT, "No users in database to remove.");
+            }
         }
     }
-    @PutMapping("/update/{id}")
-    public void updateUser(@RequestBody User u, @PathVariable("id") int id) 
+
+    @RequestMapping(value = "/update/{id}", method = RequestMethod.PUT)
+    public void updateUser(@RequestBody User u,
+                           @PathVariable("id") int id,
+                           @RequestParam(name = "sessid", required = true) String sessid)
     {
-        try
+        if (sessid.isEmpty())
         {
-            User oldUser=users.findById(id);
-            
-            oldUser.setEmail(u.getEmail());
-            oldUser.setUsername(u.getUsername());
-            oldUser.setFirstname(u.getFirstname());
-            oldUser.setLastname(u.getLastname());
-            oldUser.setPassword(u.getPassword());
-            oldUser.setUniversity(u.getUniversity());
-            log.info("Data Entry Successful: User with ID " + id + " updated.");
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Request Invalid: Empty value for required parameter 'sessid'.");
         }
-        catch(Exception e)
+        Session active = sessions.findBySessId(sessid);
+        if (active == null) throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Could not find an active session with id: " + sessid);
+        
+        if (active.getUser() == users.findById(id)
+         || active.getUser().getAdmin())
         {
-            log.error(e.getMessage());
-            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Could not update the user with id: " + id);
+            try
+            {
+                // this method of updating the user doesn't actually work. doesn't send the result to the server
+                User oldUser=users.findById(id);
+
+                oldUser.setEmail(u.getEmail());
+                oldUser.setUsername(u.getUsername());
+                oldUser.setFirstname(u.getFirstname());
+                oldUser.setLastname(u.getLastname());
+                oldUser.setPassword(u.getPassword());
+                oldUser.setUniversity(u.getUniversity());
+
+                // need something here to send the change to SQL
+                log.info("Data Entry Successful: User with ID " + id + " updated.");
+
+                // Exit
+                return;
+            }
+            catch(Exception e)
+            {
+                log.error(e.getMessage());
+                throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Could not update the user with id: " + id);
+            }
         }
     }
     @RequestMapping(value = "/toString/{id}")
